Implemented player dependent clears, momentum, and falling rows.

+++  magneticenigmatic/Main/Momentum.pde
@@ -29,7 +29,8 @@ class Momentum {
     return y;
   }
   
-  public double increaseMomentum(int x) {
+  public double increaseMomentum(double x) 
+  {
     v += x;
     return v;
   }
